refactor: separated read indicators display functions from button display script into a new script
