+++ Assets/Scripts/Topic Discussion 1/Navigator and Displays/ButtonsDisplay.cs
@@ -7,9 +7,6 @@ public class ButtonsDisplay : MonoBehaviour
     [SerializeField] private PagePrevNextButton nextPageButton;
     [SerializeField] private SectorPrevNextButton prevSectorButton;
     [SerializeField] private SectorPrevNextButton nextSectorButton;
-    [Header("Read Indicator Buttons")]
-    [SerializeField] private ReadIndicatorButton markAsReadButton;
-    [SerializeField] private ReadIndicatorButton markAsNotReadButton;
 
     private void OnEnable()
     {
@@ -133,24 +130,6 @@ public class ButtonsDisplay : MonoBehaviour
     {
         // Sets the previous sector's button title into the next sector's title
         nextSectorButton.sectorButtonText.text = nextSectorTitle;
-    }
-    #endregion
-
-    #region Read Indicator Buttons
-    public void ChangeReadIndicatorButtonsState(int currentSectorIndex, int currentPageIndex, DiscussionNavigator discNav)
-    {
-        if (!discNav.CurrentPageIsMarkedRead(currentSectorIndex, currentPageIndex))
-        {
-            // Activate the mark as read button
-            markAsReadButton.gameObject.SetActive(true);
-            markAsNotReadButton.gameObject.SetActive(false);
-        }
-        else
-        {
-            // Activate the mark as not yet read button
-            markAsReadButton.gameObject.SetActive(false);
-            markAsNotReadButton.gameObject.SetActive(true);
-        }
     }
     #endregion
 }
